Safari-611.1.10/LayoutTests/plugins/private-browsing-mode-expected.txt

Tests that the private browsing API implementation works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS plugin.privateBrowsingEnabled is true
PASS plugin.cachedPrivateBrowsingEnabled is true
PASS successfullyParsed is true

TEST COMPLETE
